requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.7.2
async-timeout==5.0.1
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-filter==24.3
django-ipware==7.0.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-pwa==2.0.1
django-six==1.0.5
django-thread==0.0.1
django_extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
frozenlist==1.5.0
gunicorn==23.0.0
http_ece==1.2.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isoweek==1.3.3
jdatetime==4.1.1
kavenegar==1.1.2
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
persiantools==3.0.1
phonenumbers==8.13.31
pillow==10.2.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydub==0.25.1
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==24.0.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==3.0.0
pytz==2024.1
pywebpush==2.0.3
redis==5.2.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
screen==1.0.1
service-identity==24.1.0
six==1.16.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
sqlparse==0.4.4
TimeConvert==3.0.13
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlwt==1.3.0
yarl==1.18.3
zope.interface==6.2
onvif-zeep==0.2.12
pyaml==25.1.0
django-celery-beat==2.7.0
django-minio-storage==0.5.8
